GUI and Handler
Now changed it so once u remove items from the order ready it upadates
the inventory scroll view using the ingredient array hardcoded in
MenuItem. Also if u log out of kitchen staff and then log back in the
updated values still stay the same once u log back in.

Will add more functionality later tomorrow or so so that we can use the
Queues defined in the handler.