KS Update
 Inventory Updates as orders are readied.
Now if not enough stock is available the item in the order queue is
changed to waiting. Also that menu items valid bit is changed to 0 so
it won’t appear on the customers screen. Finally a function checks to
see if the inventory has been updated and changes the waiting items to
not ready if items can be made.

Communicator is integrated.